requirements.txt
playwright
beautifulsoup4
lxml
fake-useragent
tldextract
pydantic
python-dotenv
Flask
2captcha-python
playwright-stealth
extruct
langdetect
requests
aiohttp
Pillow
Wappalyzer-python
dnspython
shodan
